% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{arguments}
\alias{arguments}
\title{Returns the argument list given to the current function}
\usage{
arguments()
}
\value{
The argument list
}
\description{
Returns the argument list given to the current function
}
