% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_signature}
\alias{check_signature}
\title{Check the signature of a function.}
\usage{
check_signature(formals, ..., .args = NULL, .force = FALSE,
  .allow_defaults = FALSE)
}
\arguments{
\item{formals}{A list of formal arguments to match.}

\item{...}{Arguments to match against \code{formals}.}

\item{.args}{Optional list of arguments to match against \code{formals}.}

\item{.force}{Logical; if \code{TRUE}, forces evaluation of all aruments in
\code{...} and \code{.args}.}

\item{.allow_defaults}{Logical; if \code{TRUE}, does not check for missing
values when defaults are present in \code{formals}. Not implemented yet.}
}
\value{
A named vector containing correctly ordered arguments from \code{...}
and \code{.args}. Throws an error if this is not possible to do.
}
\description{
Check the signature of a function.
}
