% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{coef.straussR}
\alias{coef.straussR}
\title{Extract parameter estimates from a \code{straussR} object.}
\usage{
\method{coef}{straussR}(object, summary = mean, show_effects = FALSE,
  unlist = TRUE)
}
\arguments{
\item{object}{A \code{straussR} object.}

\item{summary}{A function for summarizing the posterior samples of the
object. Defaults to \code{mean}.}

\item{show_effects}{Logical; If \code{TRUE}, shows the random effects.}

\item{unlist}{Logical; If \code{TRUE}, unlists each parameter summary.}
}
\value{
A list of vectors, one for each formula.
}
\description{
Extract parameter estimates from a \code{straussR} object.
}
