% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{DoubleTruncation}
\alias{DoubleTruncation}
\alias{dttruncnorm}
\alias{rdtruncnorm}
\alias{rdtruncfnorm}
\title{Double truncation of normal and folded normal distribution.}
\usage{
dttruncnorm(x, mean = 0, sd = 1, a = 1, b = -1)

rdtruncnorm(n, mean = 0, sd = 1, a = 1, b = -1)

rdtruncfnorm(n, mean = 0, sd = 1, a = 1, b = -1)
}
\description{
Double truncation of normal and folded normal distribution.
}
