% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_functions.R
\name{dppredictive}
\alias{dppredictive}
\title{Posterior predictive distribution of a \code{snifmar} object.}
\usage{
dppredictive(x, object)
}
\arguments{
\item{x}{Numeric. A vector of points where the posterior predictive
distribution is to be evaluated.}

\item{object}{A \code{snifmar} object.}
}
\value{
A vector of densities with the same arity as \code{x}.
}
\description{
Posterior predictive distribution of a \code{snifmar} object.
}
