% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dtruncnorm}
\alias{dtruncnorm}
\title{Density of the truncated normal.}
\usage{
dtruncnorm(x, mean, sd, a = -Inf, b = Inf, log = FALSE)
}
\description{
A simple wrapper around truncnorm::dtruncnorm. In contrast to
truncnorm::dtruncnorm, this one supports the \code{log} argument.
}
