% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_functions.R
\name{effect_distribution}
\alias{effect_distribution}
\title{Posterior effect size for a \code{snifmar} object.}
\usage{
effect_distribution(object, alpha = 1.96, indices = "all",
  sample_n = FALSE)
}
\arguments{
\item{object}{A \code{snifmar} object.}
}
\description{
Posterior effect size for a \code{snifmar} object.
}
