% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extract_arguments}
\alias{extract_arguments}
\title{Get the argument list from a \code{call}.}
\usage{
extract_arguments(call)
}
\arguments{
\item{call}{An object of class \code{call}.}
}
\value{
The named argument list from the \code{call}.
}
\description{
Get the argument list from a \code{call}.
}
