% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extract_function_name}
\alias{extract_function_name}
\title{Get the function \code{name} from a \code{call}.}
\usage{
extract_function_name(call)
}
\arguments{
\item{call}{An object of class \code{call}.}
}
\value{
The function of the call \code{call}, a \code{name} object.
}
\description{
Get the function \code{name} from a \code{call}.
}
