% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_to_matrix.R
\name{family_formula_to_list2}
\alias{family_formula_to_list2}
\title{Convert a family formula}
\usage{
family_formula_to_list2(formula)
}
\arguments{
\item{formula}{A formula object.}
}
\value{
A list.
}
\description{
Convert a family formula into a list containing the response, the
formula and link for the mean and standard deviation, and the prior for
the additional parameter(s). [Not implemented yet.]
}
