% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{FoldedNormal}
\alias{FoldedNormal}
\alias{dfnorm}
\alias{pfnorm}
\alias{rfnorm}
\alias{efnorm}
\alias{vfnorm}
\alias{sdfnorm}
\title{The Folded Normal Distribution}
\usage{
dfnorm(x, mean = 0, sd = 1, log = FALSE)

pfnorm(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

rfnorm(n, mean = 0, sd = 1)

efnorm(mean = 0, sd = 1)

vfnorm(mean = 0, sd = 1)

sdfnorm(mean = 0, sd = 1)
}
\arguments{
\item{x, q}{Numeric vector of quantiles.}

\item{mean}{Numeric vector of means.}

\item{sd}{Numeric vector of standard deviations.}

\item{log, log.p}{Logical; If \code{TRUE}, probabilities \code{p} are given as
\code{log(p)}.}

\item{lower.tail}{Logical; If \code{TRUE}, probabilities are
\eqn{P(X \leq x)}, otherwise \eqn{P(X \leq x)}.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length of the
vector is used.}

\item{p}{Numeric vector of probabilities.}
}
\description{
Density, distribution function, quantile function, random generation, mean
function, variance function, and standard deviation function for the folded
normal distribution with underlying mean equal to \code{mean} and underlying
standard deviation equal to \code{sd}.
}
