% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_massage.R
\name{formula_lhs_list}
\alias{formula_lhs_list}
\title{Convenience functions for formulas.}
\usage{
formula_lhs_list(formula)
}
\arguments{
\item{formula}{A formula object}
}
\description{
Convenience functions for formulas.
}
\details{
The function \code{formula_lhs_list} identifies the integer code and
    name of the link function, and the variable of a formula and returns them
    in a named list.
}
