% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formula_to_link}
\alias{formula_to_link}
\title{Matches the left-hand side of a formula object with a link.}
\usage{
formula_to_link(formula, type = "mean")
}
\arguments{
\item{formula}{A formula object.}

\item{type}{The string used as a function argument.}
}
\value{
A string, identifying a link in the 'link_list' list.
}
\description{
Matches the left-hand side of a formula object with a link.
}
