% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formula_to_list}
\alias{formula_to_list}
\title{Link-formula to link name and formula}
\usage{
formula_to_list(formula, type = "mean")
}
\value{
List of link name and formula
}
\description{
Link-formula to link name and formula
}
