% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_two_sided}
\alias{is_two_sided}
\title{Is a formula tow-sided?}
\usage{
is_two_sided(formula)
}
\arguments{
\item{formula}{A formula.}
}
\value{
Logical; \code{TRUE} if the formula is two-sided, an error otherwise.
}
\description{
Is a formula tow-sided?
}
