% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_to_matrix.R
\name{link_formula_to_list}
\alias{link_formula_to_list}
\title{Convert a link-formula to a list.}
\usage{
link_formula_to_list(formula)
}
\value{
List.
}
\description{
Convert a link-formula to a list.
}
