% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_alist}
\alias{make_alist}
\title{Make an \code{alist} with the specified \code{names}.}
\usage{
make_alist(names)
}
\arguments{
\item{names}{A list or character vector with argument names.}
}
\value{
An \code{alist} containing named elements decided by \code{names}.
}
\description{
Make an \code{alist} with the specified \code{names}.
}
