% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_massage.R
\name{massage_data}
\alias{massage_data}
\title{Massages \code{formula}, \code{priors}, and \code{data} for \code{straussR}.}
\usage{
massage_data(formula, priors, data = NULL)
}
\arguments{
\item{formula}{A formula of the form given to \code{straussR}.}

\item{priors}{A list of priors. Should match the parameters of
\code{formula}.}

\item{data}{An optional data frame.}
}
\value{
A list containing all the data needed for \code{straussR}-
}
\description{
Converts the information contained in \code{formula}, \code{priors},
and \code{data} for \code{straussR} to a form useable by \code{straussR}.
Checks for compatibility between priors and formulas and does some error
checking.
}
