% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_massage.R
\name{massage_prior}
\alias{massage_prior}
\title{Transforms a prior formula to a list.}
\usage{
massage_prior(prior, data = NULL, .force = TRUE)
}
\arguments{
\item{data}{Optional data frame.}

\item{.force}{Logical; if \code{TRUE}, forces the arguments given to the
prior.}

\item{formula}{Two-sided formula with a covariate on the left-hand side and
a prior from \code{prior_list} on the right-hand side.}
}
\value{
A list containing list containing the name of the covariate, the name
of the prior,and the values of the prior parameters. If the right hand side
refers to a factor, the expanded factor names are used instead. (Intercept)
is not included: It must be named individually.
}
\description{
Transforms a prior formula to a list.
}
