% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_massage.R
\name{massage_priors}
\alias{massage_priors}
\title{Transforms a list of priors to matrix form.}
\usage{
massage_priors(prior, data = NULL)
}
\arguments{
\item{data}{An optional data frame.}

\item{priors}{List of priors.}
}
\value{
A list containing a matrix of prior parameters and a vector of
prior types.
}
\description{
Transforms a list of priors to matrix form.
}
