% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{match_formals}
\alias{match_formals}
\title{Match formals to a list of arugments}
\usage{
match_formals(formals, ..., .args = NULL, .force = TRUE)
}
\arguments{
\item{formals}{A list of formal arguments to match.}

\item{...}{Arguments to match against \code{formals}.}

\item{.args}{Optional list of arguments to match against \code{formals}.}

\item{.force}{Logical; If \code{TRUE}, forces evaluation of the arguments.}
}
\value{
A unlisted list containing correctly ordered arguments from
\code{...} and \code{.args}. Throws an error if this is not possible to do.
}
\description{
Given a list \code{formals} of function arguments, check whether the
arguments provided in \code{...} and matches the signature given by the list.
If it matches, return the arguments in the same order as formals. If it does
not match, raise an exception.
}
