% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_models.R
\name{pcurve}
\alias{pcurve}
\title{Fixed and mixed effects meta-analysis under selection for significance}
\usage{
pcurve(d, n, c = rep(stats::qnorm(0.975), length(n)), type = "mixed")
}
\arguments{
\item{d}{Vector of nbserved effect sizes.}

\item{n}{Integer vector of sample sizes.}

\item{c}{Vector of lower cut-off values.}

\item{type}{String specifying the type of analysis.}
}
\value{
A vector of estimates and an estimate of its asymptotic covariance
   matrix. For the fixed effects model, the estimate is the estimated mean.
   For the mixed effects model, the estimates are the mean and the variance
   of the normal effect size distribution.
}
\description{
Fixed and mixed effects meta-analysis under selection for significance
}
