% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_functions.R
\name{plot.strauss}
\alias{plot.strauss}
\title{Generic plot for \code{strauss} objects.}
\usage{
\method{plot}{strauss}(obj, kind = "small_effects", options, ...)
}
\arguments{
\item{obj}{A \code{strauss} object.}

\item{kind}{String. The kind of plot to make, see \code{details} for more.}

\item{options}{List. Options used by \code{kind}.}

\item{...}{Passed to \code{plot.default}.}
}
\description{
Generic plot for \code{strauss} objects.
}
