% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_functions.R
\name{power_distribution}
\alias{power_distribution}
\title{Find the power distribution for a \code{snifmar} object.}
\usage{
power_distribution(object, n, alpha = 1.96, indices = "all",
  sample_n = FALSE)
}
\arguments{
\item{object}{A \code{snifmar} object.}
}
\description{
Find the power distribution for a \code{snifmar} object.
}
