% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_massage.R
\name{prior_list_massage}
\alias{prior_list_massage}
\title{Massages priors into triple matrix form.}
\usage{
prior_list_massage(priors)
}
\arguments{
\item{prior_list}{A list of priors.}
}
\value{
A list of nine matrices.
}
\description{
Massages priors into triple matrix form.
}
