% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prior_massage}
\alias{prior_massage}
\title{Transforms a list of priors to matrix form.}
\usage{
prior_massage(priors)
}
\arguments{
\item{priors}{List of priors.}
}
\value{
A list containing a matrix of prior parameters and a vector of
prior types.
}
\description{
Transforms a list of priors to matrix form.
}
