% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanlm.R
\name{stanlm}
\alias{stanlm}
\title{Linear regression with STAN.}
\usage{
stanlm(formula, data, p_formula, mean_priors = NULL, sd_priors = NULL,
  p_priors = NULL)
}
\arguments{
\item{formula}{The formula.}

\item{data}{The data.}

\item{p_formula}{Formula for p.}

\item{mean_priors}{List of formulas. Priors for the coefficients for
\code{mean}.}

\item{sd_priors}{List of formulas. Priors for the coefficients for
\code{sd}.}

\item{p_priors}{List of formulas. Priors for the coefficients for
\code{p}.}
}
\value{
A \code{STAN} object.
}
\description{
Linear regression with STAN with some options.
}
