% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straussR-package.R
\docType{package}
\name{straussR-package}
\alias{straussR-package}
\alias{straussR}
\title{The 'straussR' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
