% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straussR.R
\name{straussR}
\alias{straussR}
\title{Mixture-based Bayesian Meta-analysis}
\usage{
straussR(formula, priors, data, ...)
}
\arguments{
\item{formula}{Model formula. Should have the \code{z} values on the left
hand side and the random effects distribution on the right hand side.}

\item{priors}{List of formulas. Should contain priors for all parameters
mentioned in \code{formula}.}

\item{data}{An optional data frame. NOTE: Currently required.}
}
\value{
A \code{STAN} object.
}
\description{
\code{straussR} is used to obtain posteriors for mixture-based Bayesian
meta-analysis. Fixed effects, random effects, and meta-regressions are
supported.
}
