% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{TruncatedFoldedNormal}
\alias{TruncatedFoldedNormal}
\alias{dtruncfnorm}
\alias{rtruncfnorm}
\title{The Truncated Folded Normal Distribution}
\usage{
dtruncfnorm(x, mean = 0, sd = 1, a = 0, b = Inf, log = FALSE)

rtruncfnorm(n, mean = 0, sd = 1, a = -Inf, b = Inf)
}
\arguments{
\item{x, q}{Numeric vector of quantiles.}

\item{mean}{Numeric vector of means.}

\item{sd}{Numeric vector of standard deviations.}

\item{a}{Numeric vector of lower bounds. May be \code{-Inf}.}

\item{b}{Numeric Vector of upper bounds. May be \code{Inf}.}

\item{log, log.p}{Logical; If \code{TRUE}, probabilities \code{p} are given as
\code{log(p)}.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length of the
vector is used.}

\item{p}{Numeric vector of probabilities.}

\item{lower.tail}{Logical; If \code{TRUE}, probabilities are
\eqn{P(X \leq x)}, otherwise \eqn{P(X \leq x)}.}
}
\description{
Density, distribution function, quantile function, and random generation for
the truncated folded normal distribution with underlying mean equal to
\code{mean}, underlying standard deviation equal to \code{sd}, lower bound
\code{a}, and upper bound \code{b}.
}
