% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_to_matrix.R
\name{formula_labels}
\alias{formula_labels}
\title{Get term labels from a formula, intercept included.}
\usage{
formula_labels(formula)
}
\description{
Get term labels from a formula, intercept included.
}
