% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_massage.R
\name{model_matrix}
\alias{model_matrix}
\title{Convert a collection of formulas to a model matrix and a matrix of indices.}
\usage{
model_matrix(formula, priors, data = NULL)
}
\arguments{
\item{formulas}{A list of formulas.}
}
\value{
A list containing a model matrices and a matrices of indices.
}
\description{
Convert a collection of formulas to a model matrix and a matrix of indices.
}
